requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.8
Flask==2.2.2
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
nltk==3.6.5
numpy==1.24.3
packaging==23.1
pandas==2.2.3
plotly==5.11.0
polars==0.17.11
psycopg2-binary==2.9.1
pyarrow==19.0.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
pytz==2025.2
regex==2024.11.6
requests==2.31.0
six==1.17.0
SQLAlchemy==1.4.46
tenacity==9.1.2
tqdm==4.67.1
tzdata==2025.2
urllib3==1.26.18
Werkzeug==2.2.2
gunicorn==20.1.0
finnhub-python==2.4.12
pytest==7.0.0
pytest-cov==2.12.1
scikit-learn==1.1.3
kaleido==0.2.1
Flask-WTF==1.0.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
alembic==1.7.3
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
black==23.3.0
flake8==6.0.0
isort==5.9.3
pre-commit==2.15.0
yfinance==0.2.35
